Add glossary implementation
Glossary entries can be defined with `~KEY` for the
short form and `~~KEY` for the long form.
If a glossary entry is referenced for the first time it will always be
rendered as the long form.
Glossary entries can be defined in a toml file (default is glossary.toml)
similar to bibliography.

Signed-off-by: trivernis <trivernis@protonmail.com>